Move Open Graph images to respective directories
This way, they are MUCH easier to manage. Making them work for the pages
is a bit tricky, but a minor adjustment of the API helper makes it work.
With this, only content directories that contain a file called
`index.mdx` are counted as real pages. The other pages still exist for
the API route, so we can get the `og-image.jpg` files from there.