Replace Video with YouTube- and Vimeo-previews
Even the nocookie-version of YouTube embeds a cookie once a visitor
starts playing a video. I need to show a cookie notice because of that,
which is terrible.

Instead, I now only show a blurred thumbnail and add a link to YouTube
or Vimeo. If the videos are not embedded on my site, they can’t set any
cookies.